// src/biodiversity/kingdom.enum.ts
export enum KingdomType {
  Animalia = 'Animalia',
  Chromista = 'Chromista',
  Fungi = 'Fungi',
  Plantae = 'Plantae',
  Protozoa = 'Protozoa'
}

export enum KingdomTypeUrl {
  Animalia = 'http://creativesafari.com/wp-content/uploads/2015/04/Animalia-kickoff.png',
  Chromista = 'https://upload.wikimedia.org/wikipedia/commons/thumb/b/bf/Phaeodactylum_tricornutum.png/180px-Phaeodactylum_tricornutum.png',
  Fungi = 'https://upload.wikimedia.org/wikipedia/commons/4/4b/Mushrooms_in_Woodland_along_The_Avenue_-_geograph.org.uk_-_27386.jpg',
  Plantae = 'https://2.bp.blogspot.com/--occS--kl30/UihAxqmz6tI/AAAAAAAAClY/m-LoyEVjpAk/s1600/th%5B8%5D.jpg',
  Protozoa = 'https://i.ytimg.com/vi/q_owLgg3Fm0/maxresdefault.jpg',
  default = 'https://img.icons8.com/plasticine/2x/null-symbol.png'
}